+++ src/any_tree.rs
@@ -18,80 +18,3 @@ pub enum AnyTree {
     /// Key-value separated LSM-tree, see [`BlobTree`]
     Blob(BlobTree),
 }
-
-/// Unified ingestion builder over `AnyTree`
-// Keep zero allocations and direct dispatch; boxing introduces heap indirection and `dyn` adds virtual dispatch.
-// Ingestion calls use `&mut self` in tight loops; the active variant is stable and branch prediction makes the match cheap.
-// Allowing this lint preserves hot-path performance at the cost of a larger enum size.
-#[expect(clippy::large_enum_variant)]
-pub enum AnyIngestion<'a> {
-    /// Ingestion for a standard LSM-tree
-    Standard(Ingestion<'a>),
-    /// Ingestion for a [`BlobTree`] with KV separation
-    Blob(BlobIngestion<'a>),
-}
-
-impl AnyIngestion<'_> {
-    #[must_use]
-    /// Sets the sequence number used for subsequent writes
-    pub fn with_seqno(self, seqno: SeqNo) -> Self {
-        match self {
-            Self::Standard(i) => Self::Standard(i.with_seqno(seqno)),
-            Self::Blob(b) => Self::Blob(b.with_seqno(seqno)),
-        }
-    }
-
-    /// Writes a key-value pair.
-    ///
-    /// # Errors
-    ///
-    /// Will return `Err` if an IO error occurs.
-    pub fn write<K: Into<UserKey>, V: Into<UserValue>>(
-        &mut self,
-        key: K,
-        value: V,
-    ) -> crate::Result<()> {
-        match self {
-            Self::Standard(i) => i.write(key.into(), value.into()),
-            Self::Blob(b) => b.write(key.into(), value.into()),
-        }
-    }
-
-    /// Writes a tombstone for a key.
-    ///
-    /// # Errors
-    ///
-    /// Will return `Err` if an IO error occurs.
-    pub fn write_tombstone<K: Into<UserKey>>(&mut self, key: K) -> crate::Result<()> {
-        match self {
-            Self::Standard(i) => i.write_tombstone(key.into()),
-            Self::Blob(b) => b.write_tombstone(key.into()),
-        }
-    }
-
-    /// Finalizes ingestion and registers created tables (and blob files if present).
-    ///
-    /// # Errors
-    ///
-    /// Will return `Err` if an IO error occurs.
-    pub fn finish(self) -> crate::Result<()> {
-        match self {
-            Self::Standard(i) => i.finish(),
-            Self::Blob(b) => b.finish(),
-        }
-    }
-}
-
-impl AnyTree {
-    /// Starts an ingestion for any tree type (standard or blob).
-    ///
-    /// # Errors
-    ///
-    /// Will return `Err` if an IO error occurs.
-    pub fn ingestion(&self) -> crate::Result<AnyIngestion<'_>> {
-        match self {
-            Self::Standard(t) => Ok(AnyIngestion::Standard(Ingestion::new(t)?)),
-            Self::Blob(b) => Ok(AnyIngestion::Blob(BlobIngestion::new(b)?)),
-        }
-    }
-}
