+++ tests/tree_bulk_ingest.rs
@@ -13,14 +13,12 @@ fn tree_bulk_ingest() -> lsm_tree::Result<()> {
     let tree = Config::new(folder, seqno.clone()).open()?;
 
     let mut ingestion = tree.ingestion()?;
-    let seq = seqno.next();
-    ingestion = ingestion.with_seqno(seq);
     for x in 0..ITEM_COUNT as u64 {
         let k = x.to_be_bytes();
         let v = nanoid::nanoid!();
         ingestion.write(k, v)?;
     }
-    ingestion.finish()?;
+    let seq = ingestion.finish()?;
     visible_seqno.fetch_max(seq + 1);
 
     assert_eq!(tree.len(SeqNo::MAX, None)?, ITEM_COUNT);
@@ -49,14 +47,12 @@ fn tree_copy() -> lsm_tree::Result<()> {
     let src = Config::new(folder, seqno.clone()).open()?;
 
     let mut ingestion = src.ingestion()?;
-    let seq = seqno.next();
-    ingestion = ingestion.with_seqno(seq);
     for x in 0..ITEM_COUNT as u64 {
         let k = x.to_be_bytes();
         let v = nanoid::nanoid!();
         ingestion.write(k, v)?;
     }
-    ingestion.finish()?;
+    let seq = ingestion.finish()?;
     visible_seqno.fetch_max(seq + 1);
 
     assert_eq!(src.len(SeqNo::MAX, None)?, ITEM_COUNT);
@@ -76,13 +72,11 @@ fn tree_copy() -> lsm_tree::Result<()> {
     let dest = Config::new(folder, seqno.clone()).open()?;
 
     let mut ingestion = dest.ingestion()?;
-    let seq = seqno.next();
-    ingestion = ingestion.with_seqno(seq);
     for item in src.iter(SeqNo::MAX, None) {
         let (k, v) = item.into_inner().unwrap();
         ingestion.write(k, v)?;
     }
-    ingestion.finish()?;
+    let seq = ingestion.finish()?;
     visible_seqno.fetch_max(seq + 1);
 
     assert_eq!(dest.len(SeqNo::MAX, None)?, ITEM_COUNT);
@@ -113,14 +107,12 @@ fn blob_tree_bulk_ingest() -> lsm_tree::Result<()> {
         .open()?;
 
     let mut ingestion = tree.ingestion()?;
-    let seq = seqno.next();
-    ingestion = ingestion.with_seqno(seq);
     for x in 0..ITEM_COUNT as u64 {
         let k = x.to_be_bytes();
         let v = nanoid::nanoid!();
         ingestion.write(k, v)?;
     }
-    ingestion.finish()?;
+    let seq = ingestion.finish()?;
     visible_seqno.fetch_max(seq + 1);
 
     assert_eq!(tree.len(SeqNo::MAX, None)?, ITEM_COUNT);
@@ -152,14 +144,12 @@ fn blob_tree_copy() -> lsm_tree::Result<()> {
         .open()?;
 
     let mut ingestion = src.ingestion()?;
-    let seq = seqno.next();
-    ingestion = ingestion.with_seqno(seq);
     for x in 0..ITEM_COUNT as u64 {
         let k = x.to_be_bytes();
         let v = nanoid::nanoid!();
         ingestion.write(k, v)?;
     }
-    ingestion.finish()?;
+    let seq = ingestion.finish()?;
     visible_seqno.fetch_max(seq + 1);
 
     assert_eq!(src.len(SeqNo::MAX, None)?, ITEM_COUNT);
@@ -182,13 +172,11 @@ fn blob_tree_copy() -> lsm_tree::Result<()> {
         .open()?;
 
     let mut ingestion = dest.ingestion()?;
-    let seq = seqno.next();
-    ingestion = ingestion.with_seqno(seq);
     for item in src.iter(SeqNo::MAX, None) {
         let (k, v) = item.into_inner().unwrap();
         ingestion.write(k, v)?;
     }
-    ingestion.finish()?;
+    let seq = ingestion.finish()?;
     visible_seqno.fetch_max(seq + 1);
 
     assert_eq!(dest.len(SeqNo::MAX, None)?, ITEM_COUNT);
